Separate "Application" schema
