+++ innopoints/schemas/__init__.py
@@ -1,5 +1,6 @@
 from .project import *
 from .account import *
 from .activity import *
+from .application import *
 
 from .schemas import *  # temporary until moved all schemas
