+++ innopoints/schemas/schemas.py
@@ -5,10 +5,7 @@ from innopoints.extensions import ma, db
 from marshmallow_enum import EnumField
 
 from innopoints.models import (
-    Application,
-    ApplicationStatus,
     Color,
-    Competence,
     Product,
     ProductImage,
     Size,
@@ -22,19 +19,6 @@ from innopoints.models import (
 # pylint: disable=no-member
 
 
-
-
-
-class ApplicationSchema(ma.ModelSchema):
-    class Meta:
-        model = Application
-        ordered = True
-        sqla_session = db.session
-        exclude = ('report', 'feedback')
-
-    status = EnumField(ApplicationStatus)
-    applicant = ma.Nested('AccountSchema', only=('full_name', 'email'))
-    telegram_username = ma.Str(data_key='telegram')
